StudyBuddy-recomendsystem/requirements.txt
streamlit==1.29.0
pandas==2.1.0
requests==2.31.0
google-generativeai==0.3.1
fuzzywuzzy==0.18.0